requirements.txt
onnxruntime-gpu==1.4.0
numpy==1.20.1
opencv-python==4.5.1.48
